Move ownership-related parts into an external class
